docs: Added some documentation
