+++ api/src/main/java/io/github/miniplaceholders/api/placeholder/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Placeholders Package.
+ *
+ * @since 3.0.0
+ */
+package io.github.miniplaceholders.api.placeholder;
