+++ api/src/main/java/io/github/miniplaceholders/api/resolver/AudienceTagResolver.java
@@ -1,15 +1,16 @@
 package io.github.miniplaceholders.api.resolver;
 
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import net.kyori.adventure.audience.Audience;
 import net.kyori.adventure.text.minimessage.Context;
 import net.kyori.adventure.text.minimessage.tag.Tag;
 import net.kyori.adventure.text.minimessage.tag.resolver.ArgumentQueue;
+import org.jspecify.annotations.NullMarked;
 
 /**Audience Placeholder based on an Audience */
 @FunctionalInterface
+@NullMarked
 public interface AudienceTagResolver<A extends Audience> {
     /**
      * A Tag based on an Audience
@@ -17,11 +18,8 @@ public interface AudienceTagResolver<A extends Audience> {
      * @param queue the argument queue
      * @param ctx the context
      * @return a Tag
-     * @since 1.0.0
+     * @since 3.0.0
      */
-    @Nullable Tag tag(
-        @NotNull final A audience,
-        @NotNull final ArgumentQueue queue,
-        @NotNull final Context ctx
-    );
+    @Nullable
+    Tag tag(final A audience, final ArgumentQueue queue, final Context ctx);
 }
