+++ api/src/main/java/io/github/miniplaceholders/api/resolver/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Placeholders available based on information obtained from Audiences
+ * Resolvers available based on information obtained from Audiences
  */
 package io.github.miniplaceholders.api.resolver;
