+++ api/src/main/java/io/github/miniplaceholders/api/types/Platform.java
@@ -1,13 +1,33 @@
 package io.github.miniplaceholders.api.types;
 
-/**Platforms on which MiniPlaceholders can be used */
+/**
+ * Platforms on which MiniPlaceholders can be used.
+ *
+ * @since 1.0.0
+ */
 public enum Platform {
-    /**Paper Platform */
-    PAPER,
-    /**Velocity Platform */
-    VELOCITY,
-    /**Fabric Platform */
-    FABRIC,
-    /**Sponge Platform */
-    SPONGE
+  /**
+   * Paper Platform
+   *
+   * @since 1.0.0
+   */
+  PAPER,
+  /**
+   * Velocity Platform
+   *
+   * @since 1.0.0
+   */
+  VELOCITY,
+  /**
+   * Fabric Platform
+   *
+   * @since 2.1.0
+   */
+  FABRIC,
+  /**
+   * Sponge Platform
+   *
+   * @since 2.2.0
+   */
+  SPONGE
 }
